src/requirements.txt
argcomplete==1.12.3
asgiref==3.3.4
boto3==1.17.56
botocore==1.20.56
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.4
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==4.0.1
jmespath==0.10.0
kappa==0.6.0
mysqlclient==2.0.3
pep517==0.10.0
pip-tools==6.1.0
pkg-resources==0.0.0
placebo==0.9.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.60.0
troposphere==2.7.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.4.1
